+++ data-raw/spatial.R
@@ -1,16 +1,5 @@
 # Creates the macleish_layers list
-
-library(rgdal)
 dsn <- path.expand("data-raw/macleish/")
-list.files(dsn)
-ogrListLayers(dsn)
-ogrInfo(dsn, layer = "trails")
-
-# layers <- c("Key Landmarks", "VegType_2009_20101112", "Streams", 
-            # "ChallengeCourse_Elements", "Reservoir", 
-            # "Buildings", "Wetlands",
-            # "Slope2016", "SCHOOLS_PT", "2016_MacLeish_Boundary",
-            # "Research_Poly", "Rhodes_soil", "Trails_20160328")
 
 layers <- c("landmarks", "forests", "streams",
   "challenge_courses", 
@@ -24,14 +13,23 @@ library(sf)
 macleish_layers <- lapply(layers, st_read, dsn = dsn)
 names(macleish_layers) <- layers
 
-# names(macleish_layers) <- c("landmarks", "forests", "streams",
-#                             "challenge_courses", "reservoir", 
-#                             "buildings", "wetlands",
-#                             "slopes", "schools", "boundary", 
-#                             "research", "soil", "trails")
-
 # check to make sure they are all projected the same
-sapply(macleish_layers, st_crs)
+lapply(macleish_layers, st_crs)
+
+camp_sites <- tibble::tribble(
+  ~name, ~lat, ~lon, 
+  "Group Campsite", 42.450976, -72.678154, 
+  "Remote campsite", 42.458549, -72.679581
+) %>%
+  st_as_sf(coords = c("lon", "lat")) %>%
+  st_set_crs(4326)
+
+macleish_layers[["camp_sites"]] <- camp_sites
+
+# rename forests field
+macleish_layers[["forests"]] <- macleish_layers %>%
+  purrr::pluck("forests") %>%
+  dplyr::mutate(type = Sheet1__Na)
 
 # fix the projection string
 # proj4string(macleish_layers[[10]]) <- proj4string(macleish_layers[[1]])